fix: second code audit — 11 fixes (2 bugs, 5 correctness, 4 style)
Bugs:
- main.cpp: semaphore spec violation — add waitIdle before continue when
  extent_height==0 after successful acquireNextImage (signalled semaphore
  must be consumed before reuse)
- terrain.cpp: signed integer overflow UB in hash — cast int32_t operands
  to uint32_t before multiplying by large constants

Correctness:
- main.cpp: remove dead cross-frame depth barrier (depth image transitions
  from eUndefined which discards content, making the barrier a no-op)
- main.cpp: add waitIdle after OutOfDateKHR catch to reset semaphore state
- main.cpp: stale comment referencing non-existent "pad0" field
- meshlet.hpp: stale comment "255 output limit" → "256 output limit"

Style:
- device.cpp: brace init for loop variable, direct init for set
- swapchain.cpp: C-style array → std::array
- main.cpp: C-style array → std::array
- matrix.hpp, projection.hpp: int loop counters → uint32_t with brace init

+++ src/terrain.cpp
@@ -32,7 +32,8 @@ static float valueNoise(float x, float z, uint32_t seed)
 
     // Hash function — produces pseudo-random values from integer coordinates.
     auto hash = [seed](int32_t px, int32_t pz) -> float {
-        uint32_t h{static_cast<uint32_t>(px * 374761393 + pz * 668265263 + seed * 1274126177)};
+        // Use unsigned arithmetic to avoid signed integer overflow UB.
+        uint32_t h{static_cast<uint32_t>(px) * 374761393u + static_cast<uint32_t>(pz) * 668265263u + seed * 1274126177u};
         h = (h ^ (h >> 13)) * 1103515245;
         h = h ^ (h >> 16);
         return static_cast<float>(h & 0x7FFFFFFF) / static_cast<float>(0x7FFFFFFF);
